+++ apps/explorer/test/support/fakes/one_coin_source.ex
@@ -16,6 +16,7 @@ defmodule Explorer.ExchangeRates.Source.OneCoinSource do
       last_updated: Timex.now(),
       name: "",
       market_cap_usd: Decimal.new(10_000_000),
+      tvl_usd: Decimal.new(100_500_000),
       symbol: Explorer.coin(),
       usd_value: Decimal.new(1),
       volume_24h_usd: Decimal.new(1)